+++ packages/tests/__utils__/components.ts
@@ -1,0 +1,19 @@
+import { Base } from '@studiometa/js-toolkit';
+
+export class Foo extends Base {
+  static config = {
+    name: 'Foo',
+  };
+}
+
+export class Bar extends Base {
+  static config = {
+    name: 'Bar',
+  };
+}
+
+export class Baz extends Base {
+  static config = {
+    name: 'Baz',
+  };
+}