+++ packages/tests/__utils__/index.ts
@@ -1,3 +1,4 @@
+export * from './components.js';
 export * from './h.js';
 export * from './lifecycle.js';
 export * from './mockIntersectionObserver.js';